Move the `non-suspend` memos into their own files.
